requirements.txt
fastmcp>=0.4.1
aiohttp>=3.9.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiomysql>=0.2.0
pymysql>=1.1.0